Changes pt 2 (moving stuff)
