+++ src/components/pages/landing/organizations/Hero.js
@@ -1,16 +1,17 @@
 import React, { Component } from "react";
 import { withStyles, Typography, Hidden } from "@material-ui/core";
-import LandingAppBar from "../../elements/header/LandingAppBar";
-import user from "../../../stores/user";
-import { fontFamilyBold, secondaryHex } from "../../../config/theme";
+import LandingAppBar from "../../../elements/header/LandingAppBar";
+import user from "../../../../stores/user";
+import { fontFamilyBold } from "../../../../config/theme";
 
 const styles = theme => ({
 	root: {
 		backgroundColor: "#221D27",
 		display: "flex",
 		flexDirection: "column",
-		minHeight: 267,
 		justifyContent: "center",
+
+		minHeight: 267,
 		[theme.breakpoints.down("sm")]: {
 			flexDirection: "column",
 			minHeight: 350
@@ -21,18 +22,11 @@ const styles = theme => ({
 		fontFamily: fontFamilyBold,
 		color: "#fff",
 		lineHeight: "72px",
-		// marginTop: theme.spacing.unit * 4,
 		[theme.breakpoints.down("sm")]: {
 			fontSize: theme.typography.fontSize * 2.9,
 			paddingLeft: theme.spacing.unit * 3,
 			paddingRight: theme.spacing.unit * 3
 		}
-	},
-	viewMapLinkText: {
-		fontSize: 16,
-		marginLeft: theme.spacing.unit * 1.5,
-		color: "#fff",
-		cursor: "pointer"
 	},
 	headingContainer: {
 		width: 1400,
@@ -48,31 +42,21 @@ const styles = theme => ({
 	}
 });
 
-const VenueLandingHero = props => {
+const OrgLandingHero = props => {
 	return (
 		<div className={props.classes.root}>
 			<Hidden smDown>
 				<LandingAppBar
 					isAuthenticated={user.isAuthenticated}
-					history={props.history}
+					history={history}
 				/>
 			</Hidden>
 
 			<div className={props.classes.headingContainer}>
-				<Typography variant={"display1"} className={props.classes.heading}>
-					{props.pageTitle}
-				</Typography>
-				<Typography className={props.classes.subHeading}>
-					{props.pageSubTitle}
-					{props.mapLink ? (
-						<a target="_blank" href={props.mapLink}>
-							<span className={props.classes.viewMapLinkText}>View on map</span>
-						</a>
-					) : null}
-				</Typography>
+				<Typography className={props.classes.heading}>{props.pageTitle}</Typography>
 			</div>
 		</div>
 	);
 };
 
-export default withStyles(styles)(VenueLandingHero);
+export default withStyles(styles)(OrgLandingHero);
