+++ src/components/pages/landing/organizations/Index.js
@@ -1,15 +1,17 @@
 import React, { Component } from "react";
 import { withStyles, Typography, Grid, Hidden } from "@material-ui/core";
-import { observer } from "mobx-react";
-import LandingAppBar from "../../elements/header/LandingAppBar";
-import user from "../../../stores/user";
-import CityLandingHero from "./Hero";
+import { observer } from "mobx-react/index";
+import LandingAppBar from "../../../elements/header/LandingAppBar";
+import user from "../../../../stores/user";
+import OrgLandingHero from "./Hero";
+import slugResults from "../../../../stores/slugResults";
+import Loader from "../../../elements/loaders/Loader";
+import AltResults from "../cards/AltResults";
+import notifications from "../../../../stores/notifications";
+import getUrlParam from "../../../../helpers/getUrlParam";
+import Organization from "../../admin/organizations/Organization";
 import Meta from "./Meta";
-import slugResults from "../../../stores/slugResults";
-import Loader from "../../elements/loaders/Loader";
-import AltResults from "../../elements/event/AltResults";
-import notifications from "../../../stores/notifications";
-import { fontFamilyDemiBold } from "../../../config/theme";
+import { fontFamilyDemiBold } from "../../../../config/theme";
 
 const styles = theme => ({
 	root: {},
@@ -24,10 +26,12 @@ const styles = theme => ({
 });
 
 @observer
-class CityLanding extends Component {
+class OrganizationLanding extends Component {
 	constructor(props) {
 		super(props);
-		this.state = {};
+		this.state = {
+			pageTitle: ""
+		};
 	}
 
 	componentWillMount() {
@@ -37,8 +41,6 @@ class CityLanding extends Component {
 			this.props.match.params.id
 		) {
 			const { id } = this.props.match.params;
-			// eventResults.changeFilter("city", "New York");
-			// eventResults.changeFilter("state", venueFromUrl);
 			slugResults.refreshResults(
 				id,
 				() => {},
@@ -61,7 +63,14 @@ class CityLanding extends Component {
 		return (
 			<div className={classes.root}>
 				<Meta
-					cityName={slugResults.cityInfo ? slugResults.cityInfo.city : ""}
+					cityName={
+						slugResults.orgInfo
+							? slugResults.orgInfo.city
+								? `in ${slugResults.orgInfo.city} `
+								: ""
+							: ""
+					}
+					orgName={slugResults.orgInfo ? `to ${slugResults.orgInfo.name}` : ""}
 				/>
 				<Hidden smDown>
 					<LandingAppBar
@@ -73,11 +82,11 @@ class CityLanding extends Component {
 				{slugResults.isLoading ? (
 					<Loader>Finding events...</Loader>
 				) : (
-					<CityLandingHero
+					<OrgLandingHero
 						pageTitle={
-							slugResults.cityInfo
-								? slugResults.cityInfo.city
-								: "No events for this city"
+							slugResults.orgInfo
+								? slugResults.orgInfo.name
+								: "No events for this organization"
 						}
 						history={history}
 					/>
@@ -94,7 +103,7 @@ class CityLanding extends Component {
 						) : (
 							<div>
 								<Typography variant={"display1"} className={classes.heading}>
-									Upcoming Events in {slugResults.cityInfo.city}
+									Upcoming Events at {slugResults.orgInfo.name}
 								</Typography>
 								<AltResults/>
 							</div>
@@ -106,4 +115,4 @@ class CityLanding extends Component {
 	}
 }
 
-export default withStyles(styles)(CityLanding);
+export default withStyles(styles)(OrganizationLanding);
