+++ src/components/pages/landing/venues/Hero.js
@@ -1,17 +1,16 @@
 import React, { Component } from "react";
 import { withStyles, Typography, Hidden } from "@material-ui/core";
-import LandingAppBar from "../../elements/header/LandingAppBar";
-import user from "../../../stores/user";
-import { fontFamilyBold } from "../../../config/theme";
+import LandingAppBar from "../../../elements/header/LandingAppBar";
+import user from "../../../../stores/user";
+import { fontFamilyBold, secondaryHex } from "../../../../config/theme";
 
 const styles = theme => ({
 	root: {
 		backgroundColor: "#221D27",
 		display: "flex",
 		flexDirection: "column",
-		justifyContent: "center",
-
 		minHeight: 267,
+		justifyContent: "center",
 		[theme.breakpoints.down("sm")]: {
 			flexDirection: "column",
 			minHeight: 350
@@ -22,11 +21,18 @@ const styles = theme => ({
 		fontFamily: fontFamilyBold,
 		color: "#fff",
 		lineHeight: "72px",
+		// marginTop: theme.spacing.unit * 4,
 		[theme.breakpoints.down("sm")]: {
 			fontSize: theme.typography.fontSize * 2.9,
 			paddingLeft: theme.spacing.unit * 3,
 			paddingRight: theme.spacing.unit * 3
 		}
+	},
+	viewMapLinkText: {
+		fontSize: 16,
+		marginLeft: theme.spacing.unit * 1.5,
+		color: "#fff",
+		cursor: "pointer"
 	},
 	headingContainer: {
 		width: 1400,
@@ -42,21 +48,31 @@ const styles = theme => ({
 	}
 });
 
-const OrgLandingHero = props => {
+const VenueLandingHero = props => {
 	return (
 		<div className={props.classes.root}>
 			<Hidden smDown>
 				<LandingAppBar
 					isAuthenticated={user.isAuthenticated}
-					history={history}
+					history={props.history}
 				/>
 			</Hidden>
 
 			<div className={props.classes.headingContainer}>
-				<Typography className={props.classes.heading}>{props.pageTitle}</Typography>
+				<Typography variant={"display1"} className={props.classes.heading}>
+					{props.pageTitle}
+				</Typography>
+				<Typography className={props.classes.subHeading}>
+					{props.pageSubTitle}
+					{props.mapLink ? (
+						<a target="_blank" href={props.mapLink}>
+							<span className={props.classes.viewMapLinkText}>View on map</span>
+						</a>
+					) : null}
+				</Typography>
 			</div>
 		</div>
 	);
 };
 
-export default withStyles(styles)(OrgLandingHero);
+export default withStyles(styles)(VenueLandingHero);
